% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chebyshev.R
\name{Chebyshev}
\alias{Chebyshev}
\title{Chebyshev}
\usage{
Chebyshev(k)
}
\arguments{
\item{k}{a quantitative vector}
}
\value{
A number determining if Chebyshev works
}
\description{
Chebyshev
}
\examples{
Chebyshev(1:10)
}
