% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Firedam.R
\docType{data}
\name{FIREDAM}
\alias{FIREDAM}
\title{FIREDAM Dataset}
\format{
A data frame with the distance from a fire station and estimated damage costs.
}
\source{
The data came from an insurance company
}
\usage{
data(FIREDAM)
}
\description{
This dataset contains information about distance from a fire station and estimated damage costs.
}
\examples{
data(FIREDAM)
head(FIREDAM)

}
\keyword{datasets}
