% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mybin.R
\name{mybin}
\alias{mybin}
\title{MYbin}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{n}{- size of sample}

\item{p}{- probability}

\item{iter-}{number of iterations}
}
\value{
a quantity
}
\description{
MYbin
}
