% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{gamma}{- percent no show}

\item{p}{percent show}

\item{N-number}{of seats}
}
\value{
optimal number of tickets to sell
}
\description{
ntickets
}
\examples{
ntickets(400, 0.02, 0.95)
}
